+++ src/test/java/com/my/relink/service/UserServiceTest.java
@@ -34,9 +34,6 @@ class UserServiceTest extends DummyObject {
     UserService userService;
 
     @Mock
-    ReviewRepository reviewRepository;
-
-    @Mock
     private UserRepository userRepository;
 
     @Mock
@@ -44,56 +41,6 @@ class UserServiceTest extends DummyObject {
 
     @Mock
     private ImageRepository imageRepository;
-
-
-    @Test
-    @DisplayName("유저 신뢰도 계산: 정상 케이스")
-    void getTrustScore_success(){
-        User user = mockRequesterUser();
-        when(reviewRepository.getTotalStarAvg(user)).thenReturn(4.5);
-
-        int trustScore = userService.getTrustScore(user);
-
-        assertThat(trustScore).isEqualTo(90);
-        verify(reviewRepository, times(1)).getTotalStarAvg(user);
-    }
-
-    @Test
-    @DisplayName("유저 신뢰도 점수 계산: 리뷰가 없는 경우 신뢰도는 0이어야 한다")
-    void getTrustScore_NoReviews() {
-        User user = mockRequesterUser();
-        when(reviewRepository.getTotalStarAvg(user)).thenReturn(null);
-
-        int trustScore = userService.getTrustScore(user);
-
-        assertThat(trustScore).isEqualTo(0);
-        verify(reviewRepository, times(1)).getTotalStarAvg(user);
-    }
-
-    @Test
-    @DisplayName("유저 신뢰도 점수 계산: 최대 점수는 100이어야 한다")
-    void getTrustScore_MaxScore() {
-        User user = mockRequesterUser();
-        when(reviewRepository.getTotalStarAvg(user)).thenReturn(5.0);
-
-        int trustScore = userService.getTrustScore(user);
-
-        assertThat(trustScore).isEqualTo(100);
-        verify(reviewRepository, times(1)).getTotalStarAvg(user);
-    }
-
-    @Test
-    @DisplayName("유저 신뢰도 점수 계산: 최소 점수는 10이어야 한다")
-    void getTrustScore_MinScore() {
-        User user = mockRequesterUser();
-        when(reviewRepository.getTotalStarAvg(user)).thenReturn(0.5);
-
-        int trustScore = userService.getTrustScore(user);
-
-        assertThat(trustScore).isEqualTo(10);
-        verify(reviewRepository, times(1)).getTotalStarAvg(user);
-    }
-
 
 
     @Test
